requirements.txt
mysqlclient
sqlalchemy
records
discord.py
